+++ command.d.ts
@@ -1,0 +1,6 @@
+declare global {
+  interface HTMLElementTagNameMap {
+    'm-command-palette': HTMLElement & { refresh(): void };
+  }
+}
+export {};